fix: app open bug
Fix app shortcuts bug

Make app shortcuts follow tile logic and use recent tunnel, then primary, then first if tunnel name not specified
