+++ app/src/main/res/xml/shortcuts.xml
@@ -4,7 +4,7 @@
         android:icon="@drawable/vpn_on"
         android:shortcutDisabledMessage="@string/vpn_on"
         android:shortcutId="defaultOn1"
-        android:shortcutLongLabel="@string/default_vpn_on"
+        android:shortcutLongLabel="@string/vpn_on"
         android:shortcutShortLabel="@string/vpn_on">
         <intent
             android:action="START"
@@ -21,7 +21,7 @@
         android:icon="@drawable/vpn_off"
         android:shortcutDisabledMessage="@string/vpn_off"
         android:shortcutId="defaultOff1"
-        android:shortcutLongLabel="@string/default_vpn_off"
+        android:shortcutLongLabel="@string/vpn_off"
         android:shortcutShortLabel="@string/vpn_off">
         <intent
             android:action="STOP"
